+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Uncommented elements needed to build `additive-manufacturing`
 * Fix typo in selector for `labs-assessments` in `computer-science`
 
 ## [v2.19.0] - 2024-06-28